Cruster 0.7.0 - cli: briefly printing
